requirements.txt
pytest==5.4.1
selenium==3.141.0
webdriver-manager==2.4.0
pyhamcrest==2.0.2